convert point of sale resolver into a class
