+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -65,6 +65,7 @@ export {
   getPersonalizedRewriteData,
   normalizePersonalizedRewrite,
   CdpHelper,
+  PosResolver,
 } from '@sitecore-jss/sitecore-jss/personalize';
 export { GraphQLRequestClient } from '@sitecore-jss/sitecore-jss';
