+++ packages/sitecore-jss/src/personalize/pos-resolver.ts
@@ -1,0 +1,9 @@
+import { SiteInfo } from '../site';
+
+export class PosResolver {
+  static resolve = (site: SiteInfo, language: string) => {
+    return site.pointOfSale
+      ? site.pointOfSale[language] || site.pointOfSale[site.language] || site.pointOfSale['*']
+      : '';
+  };
+}
